Refractor ORM models
