+++ api/src/models/test.py
@@ -1,0 +1,13 @@
+from sqlalchemy import Column, Integer, String, ForeignKey, Enum, DateTime
+from base_class import Base
+
+class Test(Base):
+    id = Column(Integer, primary_key=True)
+    name = Column(String(255), nullable=False)
+    description = Column(String(255), nullable=False)
+    created_at = Column(DateTime, nullable=False)
+    created_by = Column(Integer, ForeignKey("user.id"), primary_key=True)
+    updated_at = Column(DateTime, nullable=False)
+    updated_by = Column(Integer, ForeignKey("user.id"), primary_key=True)
+
+
